refactoring package for some classes
